+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1769] Added warning when VRCFury version < 1.1250.0 is present with Mesh Cutter or Shape Changer delete mode components, as these combinations are incompatible.
 - Added `MA Fit Preview`, a new feature which allows you to pose a copy of your avatar, with physbones moving in edit mode.
   - You can adjust constraint and physbone settings, and have them reflected in the preview copy in real time. 
+- [#1786] Implement `MA Global Collider`
 
 ### Fixed
 - [#1791] `World Fixed Object` would double the scale of objects it's attached to